lsfd: fill TYPE field
