+++ misc-utils/Makemodule.am
@@ -250,7 +250,11 @@ if BUILD_LSFD
 bin_PROGRAMS += lsfd
 # MANPAGES += misc-utils/lsfd.1
 lsfd_SOURCES = \
-	misc-utils/lsfd.c
+	misc-utils/lsfd.c \
+	misc-utils/lsfd.h \
+	misc-utils/lsfd-file.c \
+	misc-utils/lsfd-fd-file.c \
+	misc-utils/lsfd-regular-fd-file.c
 lsfd_LDADD = $(LDADD) libsmartcols.la libcommon.la
 lsfd_LDFLAGS =  -pthread
 lsfd_CFLAGS = $(AM_CFLAGS) -I$(ul_libsmartcols_incdir) -pthread
